examples/cloudml-document-understanding/requirements.txt
gcsfs
ghostscript
google-api-python-client
google-auth
google-auth-httplib2
google-cloud-automl
google-cloud-storage
google-cloud-bigquery
pandas
wand
